# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Craftive REST API",
        "description": "Dokumentasi interaktif REST API Craftive - Premium Handmade Goods Marketplace dengan integrasi Agentic AI Custom Planner. Dikembangkan sebagai Tugas Akhir Pemrograman API - UNESA 2026.",
        "contact": {
            "email": "dev@craftive.id"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Server Lokal"
        }
    ],
    "paths": {
        "/ai/recommend": {
            "post": {
                "tags": [
                    "AI Recommendation"
                ],
                "summary": "Rekomendasi Produk AI",
                "description": "Merekomendasikan hingga 3 produk sesuai budget dan style. Tidak perlu login, cukup API Key.",
                "operationId": "92e391776f9cb1fc73a7ac1f7c595cbb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "budget"
                                ],
                                "properties": {
                                    "budget": {
                                        "description": "Budget maksimum dalam Rupiah",
                                        "type": "number",
                                        "example": 500000
                                    },
                                    "style": {
                                        "description": "Gaya produk (opsional)",
                                        "type": "string",
                                        "example": "minimalist"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rekomendasi berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "CRAFTIVE AI Recommendation"
                                        },
                                        "recommendations": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "price": {
                                                        "type": "number"
                                                    },
                                                    "rating": {
                                                        "type": "number"
                                                    },
                                                    "match": {
                                                        "type": "string",
                                                        "example": "92%"
                                                    },
                                                    "image_url": {
                                                        "type": "string"
                                                    },
                                                    "style": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                }
            }
        },
        "/buyer/custom-planner": {
            "post": {
                "tags": [
                    "AI Recommendation"
                ],
                "summary": "AI Custom Planner (Kriya)",
                "description": "Merencanakan produk kriya custom. AI menghitung biaya, tingkat kesulitan, estimasi hari, dan merekomendasikan artisan. Wajib login sebagai buyer.",
                "operationId": "1a742cdcab52f00333360b72ce7f03e5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "specifications",
                                    "materials",
                                    "budget",
                                    "timeline"
                                ],
                                "properties": {
                                    "specifications": {
                                        "type": "string",
                                        "example": "Tas anyaman rotan ukuran 30x20cm dengan tali panjang"
                                    },
                                    "materials": {
                                        "type": "string",
                                        "example": "rotan alam, benang katun"
                                    },
                                    "budget": {
                                        "description": "Minimal Rp 10.000",
                                        "type": "number",
                                        "example": 750000
                                    },
                                    "timeline": {
                                        "description": "Hari pengerjaan (minimal 1)",
                                        "type": "integer",
                                        "example": 14
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rencana berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "CRAFTIVE AI Custom Planner"
                                        },
                                        "planning": {
                                            "properties": {
                                                "difficulty": {
                                                    "type": "string",
                                                    "example": "Sedang"
                                                },
                                                "material_cost": {
                                                    "type": "number",
                                                    "example": 262500
                                                },
                                                "labor_cost": {
                                                    "type": "number",
                                                    "example": 337500
                                                },
                                                "estimated_price": {
                                                    "type": "number",
                                                    "example": 600000
                                                },
                                                "estimated_days": {
                                                    "type": "integer",
                                                    "example": 13
                                                },
                                                "shop_recommendation": {
                                                    "type": "string",
                                                    "example": "Sanggar Batik Nusantara"
                                                },
                                                "agent_reasoning": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Token tidak valid"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token JWT dari endpoint Login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "basicAuth": {
                "type": "http",
                "description": "HTTP Basic Auth (email & password)",
                "scheme": "basic"
            }
        }
    },
...
}